# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  alerts.Alert:
    properties:
      acknowledged_at:
        type: string
      acknowledged_by:
        type: string
      created_at:
        type: string
      device_id:
        type: string
      device_name:
        type: string
      id:
        type: string
      message:
        type: string
      metric_value:
        type: number
      resolution_note:
        type: string
      resolved_at:
        type: string
      resolved_by:
        type: string
      severity:
        type: string
      status:
        type: string
      threshold_value:
        type: number
      title:
        type: string
      type:
        type: string
    type: object
  alerts.AlertListResponse:
    properties:
      alerts:
        items:
          $ref: '#/definitions/alerts.Alert'
        type: array
      total:
        type: integer
    type: object
  alerts.ResolveRequest:
    properties:
      resolution_note:
        type: string
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/auth.UserDTO'
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
      updated_at:
        type: string
    type: object
  auth.UserDTO:
    properties:
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      role:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Remote Monitoring & Management Platform API
  title: RMM Platform API
  version: "1.0"
paths:
  /alerts:
    get:
      description: Get paginated list of alerts with optional filters
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 50, max 100)
        in: query
        name: limit
        type: integer
      - description: Filter by status (open/acknowledged/resolved)
        in: query
        name: status
        type: string
      - description: Filter by severity (critical/warning/info)
        in: query
        name: severity
        type: string
      - description: Filter by device ID
        in: query
        name: device_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alerts.AlertListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List alerts
      tags:
      - alerts
  /alerts/{id}:
    get:
      description: Get detailed alert information
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alerts.Alert'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get alert by ID
      tags:
      - alerts
  /alerts/{id}/acknowledge:
    post:
      description: Mark alert as acknowledged
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alerts.Alert'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Acknowledge alert
      tags:
      - alerts
  /alerts/{id}/resolve:
    post:
      consumes:
      - application/json
      description: Mark alert as resolved with optional note
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: string
      - description: Resolution note
        in: body
        name: body
        schema:
          $ref: '#/definitions/alerts.ResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/alerts.Alert'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Resolve alert
      tags:
      - alerts
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, returns JWT tokens
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout user
      tags:
      - auth
  /auth/me:
    get:
      description: Get authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh JWT token
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"